Create core/url_utils.py, moving utilities from lms_xblock/runtime
