+++ lms/djangoapps/lms_xblock/runtime.py
@@ -1,8 +1,6 @@
 """
 Module implementing `xblock.runtime.Runtime` functionality for the LMS
 """
-import re
-
 from django.conf import settings
 from django.core.urlresolvers import reverse
 
@@ -10,6 +8,7 @@ from badges.service import BadgingService
 from badges.utils import badges_enabled
 from openedx.core.djangoapps.user_api.course_tag import api as user_course_tag_api
 from openedx.core.lib.xblock_utils import xblock_local_resource_url
+from openedx.core.lib.url_utils import quote_slashes
 from request_cache.middleware import RequestCache
 import xblock.reference.plugins
 from xmodule.library_tools import LibraryToolsService
@@ -19,55 +18,6 @@ from xmodule.services import SettingsService
 from xmodule.x_module import ModuleSystem
 
 from lms.djangoapps.lms_xblock.models import XBlockAsidesConfig
-
-
-def _quote_slashes(match):
-    """
-    Helper function for `quote_slashes`
-    """
-    matched = match.group(0)
-    # We have to escape ';', because that is our
-    # escape sequence identifier (otherwise, the escaping)
-    # couldn't distinguish between us adding ';_' to the string
-    # and ';_' appearing naturally in the string
-    if matched == ';':
-        return ';;'
-    elif matched == '/':
-        return ';_'
-    else:
-        return matched
-
-
-def quote_slashes(text):
-    """
-    Quote '/' characters so that they aren't visible to
-    django's url quoting, unquoting, or url regex matching.
-
-    Escapes '/'' to the sequence ';_', and ';' to the sequence
-    ';;'. By making the escape sequence fixed length, and escaping
-    identifier character ';', we are able to reverse the escaping.
-    """
-    return re.sub(ur'[;/]', _quote_slashes, text)
-
-
-def _unquote_slashes(match):
-    """
-    Helper function for `unquote_slashes`
-    """
-    matched = match.group(0)
-    if matched == ';;':
-        return ';'
-    elif matched == ';_':
-        return '/'
-    else:
-        return matched
-
-
-def unquote_slashes(text):
-    """
-    Unquote slashes quoted by `quote_slashes`
-    """
-    return re.sub(r'(;;|;_)', _unquote_slashes, text)
 
 
 def handler_url(block, handler_name, suffix='', query='', thirdparty=False):
