+++ openedx/core/lib/tests/test_xblock_utils.py
@@ -9,7 +9,7 @@ import uuid
 
 from django.test.client import RequestFactory
 
-from lms.djangoapps.lms_xblock.runtime import quote_slashes
+from openedx.core.lib.url_utils import quote_slashes
 from xblock.fragment import Fragment
 from xmodule.modulestore import ModuleStoreEnum
 from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
