+++ openedx/core/lib/url_utils.py
@@ -1,0 +1,53 @@
+"""
+Contains common utilities for URL escaping.
+"""
+import re
+
+
+def quote_slashes(text):
+    """
+    Quote '/' characters so that they aren't visible to
+    django's url quoting, unquoting, or url regex matching.
+
+    Escapes '/'' to the sequence ';_', and ';' to the sequence
+    ';;'. By making the escape sequence fixed length, and escaping
+    identifier character ';', we are able to reverse the escaping.
+    """
+    return re.sub(ur'[;/]', _quote_slashes, text)
+
+
+def unquote_slashes(text):
+    """
+    Unquote slashes quoted by `quote_slashes`
+    """
+    return re.sub(r'(;;|;_)', _unquote_slashes, text)
+
+
+def _quote_slashes(match):
+    """
+    Helper function for `quote_slashes`
+    """
+    matched = match.group(0)
+    # We have to escape ';', because that is our
+    # escape sequence identifier (otherwise, the escaping)
+    # couldn't distinguish between us adding ';_' to the string
+    # and ';_' appearing naturally in the string
+    if matched == ';':
+        return ';;'
+    elif matched == '/':
+        return ';_'
+    else:
+        return matched
+
+
+def _unquote_slashes(match):
+    """
+    Helper function for `unquote_slashes`
+    """
+    matched = match.group(0)
+    if matched == ';;':
+        return ';'
+    elif matched == ';_':
+        return '/'
+    else:
+        return matched
